refactor: commons 디렉토리 구조로 공통 모듈 재구성
- API: battlePage/apis → commons/apis
- 컴포넌트: src/components → commons/components
