+++ frontend/src/commons/apis/getBattleInfo.ts
